backend/requirements.txt
fastapi==0.110.0
uvicorn==0.24.0
SQLAlchemy==2.0.29
python-dotenv==1.0.0
cryptography==42.0.5
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2
psycopg2-binary==2.9.9
python-multipart==0.0.9